--- .env.example
@@ -1,8 +1,0 @@
-POSTGRES_DB=*******
-POSTGRES_USER=********
-POSTGRES_PASSWORD=*********
-
-NODE_ENV=development
-SOCKET_PORT=3001
-
-NEXT_PUBLIC_SOCKET_URL=http://localhost:3001
