+++ frontend/app/game/constants.ts
@@ -1,0 +1,2 @@
+export const PLAYER_RADIUS = 0.6;
+export const WORLD_SIZE = 28;
