+++ frontend/app/game/types.ts
@@ -1,0 +1,11 @@
+export type PlayerState = {
+  id: string;
+  name: string;
+  position: { x: number; y: number; z: number };
+  velocity: { x: number; z: number };
+};
+
+export type RoomStatePayload = {
+  roomId: string;
+  players: PlayerState[];
+};
